Put codemirror in the master branch as well
People having problems using js-beautifier checkout locally.

If anyone's willing to update CodeMirror to a recent version, they're
most welcome.

I'm not using git externals, as they seem to bork the github page
generation silently and completely, if used under gh-pages branch (which
is exactly what I'd need) (an up-to-date copy, not a silent complete
bork, that is).