Fix for HTML wrap_line_length
When given a wrap_line_length of 0 without specifying a max_char
then the given wrap_line_length would be overwritten by an
undefined value for max_char.
This results in the html beautifier ignoring the wrap_line_length
of 0 given by the user and setting a default wrap_line_length of
250.

Fix this by removing the check for wrap_line_length === 0 in
the max_char compatibility if condition.

A test is included to set the wrap_line_length to 0 and call
the beautifier with a p tag containing a string which is longer
than 250 characters and expecting no changes in the string.