Support SASS nested pseudo-classes and parent reference
JavaScript:
css beautifier checks if the colon it encounters
is part of a property: value or part of a nested pseudo-class/
parent reference. In case of nested pseudo-class or parent reference
it does not output a space after the colon.

Added general test for parent reference, nested pseudo-class and
SASS import statement. Added special test for issues #410 and #414.

Python:
Best effort to support parent reference and nested pseudo-class.
When it encounters a colon it checks if the colon is part
of a property: value or part of parent reference/nested pseudo-class.
In case of parent reference/nested pseudo-class it does not output
a space after the colon.

Added tests for reference/nested pseudo-class. Add tests for
issues #410, #414

Fixes issues #410 and #414
Related to issue #356