Remove Python interdependency (Fixes #1808)
Installing jsbeautifier and cssbeautifier at the same
time fails, because the cssbeautifier setup.py tries to
import jsbeautifier.__version__ (which is not installed
at that time).

As long as both are separate packages, it makes sense
for them to have their own version information (even if
they are kept in sync).

This way, the packages can be installed separately and
in any order without problems.